add container to bookmark references: https://github.com/metanorma/isodoc/issues/534
